Update into more structures with folder and add userId field to the t…

+++ restful-api-mongo/src/controllers/tasks.controller.js
@@ -1,29 +1,44 @@
 import Tasks from '../models/tasks.model.js';
+import { serverError } from '../utils/handlers.js';
 
 export const getAllTasks = async (req, res) => {
   try {
     const tasks = await Tasks.find({});
     res.json(tasks);
   } catch (err) {
-    res.status(500).send(err); 
+    serverError(err); 
+  }
+};
+
+export const getAllTasksByUser = async (req, res) => {
+  try {
+    const tasks = await Tasks.find({ userId: req.user.userId});
+    res.json(tasks);
+  } catch (err) {
+    serverError(err); 
   }
 };
 
 export const getTask = async (req, res) => {
   try {
-    const task = await Tasks.findById(req.params.taskId);
+    const userId = req.user.userId;
+    const taskId = req.params.taskId;
+
+    const task = await Tasks.findOne({ _id: taskId, userId: userId });
+
     if (!task) {
       return res.status(404).json({ message: 'Task not found' }); 
     }
     res.json(task);
   } catch (err) {
-    res.status(500).send(err);
+    serverError(err); 
   }
 };
 
 export const createTask = async (req, res) => {
   const newTask = new Tasks({
     name: req.body.name,
+    userId: req.user.userId,
   });
   try {
     const savedTask = await newTask.save();
@@ -36,7 +51,9 @@ export const createTask = async (req, res) => {
 export const updateTask = async (req, res) => {
   try {
     const updatedTask = await Tasks.findOneAndUpdate(
-      { _id: req.params.taskId },
+      { _id: req.params.taskId,
+        userId: req.user.userId
+       },
       req.body,
       { new: true } 
     );
@@ -51,12 +68,12 @@ export const updateTask = async (req, res) => {
 
 export const deleteTask = async (req, res) => {
   try {
-    const deletedTask = await Tasks.deleteOne({ _id: req.params.taskId });
+    const deletedTask = await Tasks.deleteOne({ _id: req.params.taskId, userId: req.user.userId });
      if (deletedTask.deletedCount === 0) {
         return res.status(404).json({ message: 'Task not found' });
     }
     res.status(204).send();
   } catch (err) {
-    res.status(500).send(err);
+    serverError(err); 
   }
 };
